eliminate static methods
